Silence compiler warnings.
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@21251 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/TortoiseProc/Commands/AddCommand.cpp
@@ -70,7 +70,7 @@ bool AddCommand::Execute()
                                     taskdlg.SetExpansionArea(CString(MAKEINTRESOURCE(IDS_WARN_RELOCATEREALLY_TASK5)));
                                     taskdlg.SetDefaultCommandControl(IDCUSTOM1);
                                     taskdlg.SetMainIcon(TD_WARNING_ICON);
-                                    ret = taskdlg.DoModal(GetExplorerHWND());
+                                    ret = (UINT)taskdlg.DoModal(GetExplorerHWND());
                                 }
                                 else
                                 {
